Make user_id consistently BIGINT

+++ src/db.rs
@@ -63,7 +63,7 @@ CREATE TABLE notifications (
 ",
     "
 CREATE TABLE users (
-    user_id INTEGER PRIMARY KEY,
+    user_id BIGINT PRIMARY KEY,
     username TEXT NOT NULL
 );
 ",
